fibonacci/generator/requirements.txt
pika==1.2.0
python-dotenv==0.17.1
